Update a test that changes behavior
It's actually fine for this test to change it's behavior, because in a
sense Sorbet is more accurately saying, "one of the branches of the
`case` statement doesn't evaluate to a value" (though the error message
that it presents to say that is kind of strange). The `String` type
doesn't show up in the error message anymore because the branch where
the method *does* return `String` is not ill-typed. It's the *other*
branches (i.e., the implicit `else`) that are untyped.